optibat/pyproject.toml
[build-system]
requires = ["setuptools>=80.0.1"]
build-backend = "setuptools.build_meta"

[project]
name = "optibat"
version = "5.0.0"
requires-python = ">=3.11"
dependencies = [
    "dynaconf>=3.2.10",
    "filelock>=3.18.0",
    "numpy>=2.2.6",
    "oracledb>=3.1.0",
    "pandas>=2.2.3",
    "PIconnect>=0.12.4",
    "pyomo>=6.9.2",
    "python-box>=7.3.2",
    "SQLAlchemy>=2.0.40",
    "streamlit>=1.45.0",
    "tzdata>=2025.2",
]

[project.optional-dependencies]
dev = ["ruff>=0.11.2"]

[project.scripts]
optibat = "optibat.__main__:main"

[tool.setuptools.package-data]
optibat = ["sql/**/*", "static/**/*"]
